requirements.txt
Django==6.0
pandas==2.2.0
openpyxl==3.1.2
mysqlclient==2.2.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
django-filter==24.2
drf-yasg==1.21.11
setuptools>=65.0.0
python-dotenv>=1.0.0